convert to stories.tsx
